+++ testTslint/withTslint.ts
@@ -1,0 +1,4 @@
+// Settings (tslint): single-quote, trailing-comma, no-semi
+function foo(){
+    return "bar";
+}